Matches can be edited and deleted Streamer are now fetched from the db and the hostname of the embed twitch player is adjusted dynamically Social links are now only displayed if the member has them

+++ src/main/java/de/cschillingtschuehly/gcwebx/repositories/MatchRepository.java
@@ -3,5 +3,5 @@ package de.cschillingtschuehly.gcwebx.repositories;
 import de.cschillingtschuehly.gcwebx.modell.Match;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatchRepository extends JpaRepository<Match,Long> {
+public interface MatchRepository extends JpaRepository<Match,Integer> {
 }
